requirements.txt
asgiref==3.8.1
Django==5.0.6
django-core==1.4.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2024.1
